Adding text and link to Avataaars
Adding the avataaar resources to the brand site. Feel free to reject if there was a decision made not to include these here for some reason. I could also see this being part of an "Imagery" page rather than under "Icons." I wanted to do this because it took me a little time to find the references in the 18f-branding channel and I thought others might be looking for the Avataaars where they found the icons as well.